Refactor code to use meaningful variable names and improve readability
- Renamed variables and functions to use more descriptive names
- Extracted repeated code into separate functions for better organization
- Replaced hard-coded values with variables from the config file
- Fixed formatting and indentation issues
- Removed unused code and dependencies
- Updated event handlers for message deletion and edit to include embeds
- Updated MongoDB connection setup to remove unused code and options
